requirements.txt
requests>=2.25.1
python-telegram-bot>=20.0
apify-client>=1.0.0
python-dotenv>=0.19.0 